fix(dashboard): keep widget components out of the initial bundle
The aggregate widgets barrel re-exported the per-folder barrels, which
export the component classes. provideCoreDashboardWidgets() is referenced
from the root bootstrap module, so every widget — and its chart/statistic
dependencies — landed in the INITIAL bundle: 158.82 MB against a 150 MB
budget, and the lazy loadComponent design was defeated.

The barrel and the provider now import only the component-free
*.widgets.ts config modules.
